some refactors to prepare for barrel build support
